requirements.txt
pytest
selene ~= 2.0.0rc4
selenium
allure-pytest==2.13.2
